+++ packages/plugins/content/slate/src/Controls/index.tsx
@@ -1,5 +1,4 @@
 import { lazyLoad } from '@react-page/core';
-import { BottomToolbar } from '@react-page/ui';
 import isHotkey from 'is-hotkey';
 import debounce from 'lodash.debounce';
 import isObject from 'lodash.isobject';
@@ -8,122 +7,28 @@ import React, {
   useCallback,
   useEffect,
   useMemo,
-  useRef,
   useState
 } from 'react';
-import { Portal } from 'react-portal';
-import { createEditor, Editor, Node, Transforms } from 'slate';
+import { createEditor, Node } from 'slate';
 import {
   Editable,
-  ReactEditor,
   RenderElementProps,
   RenderLeafProps,
   Slate,
   useSlate,
   withReact
 } from 'slate-react';
-import { SlatePlugin } from 'src/types/SlatePlugin';
 import { addPlugin } from '../hooks/useAddPlugin';
 import { getCurrentNodeWithPlugin } from '../hooks/useCurrentNodeWithPlugin';
 import { removePlugin } from '../hooks/useRemovePlugin';
-import HtmlToSlate from '../HtmlToSlate';
+import withInline from '../slateEnhancer/withInline';
+import withPaste from '../slateEnhancer/withPaste';
 import { SlateProps } from '../types/component';
-import {
-  PluginButtonProps,
-  SlateComponentPluginDefinition
-} from '../types/slatePluginDefinitions';
+import { SlatePlugin } from '../types/SlatePlugin';
+import { SlateComponentPluginDefinition } from '../types/slatePluginDefinitions';
 
-const PluginButton = lazyLoad(
-  () =>
-    (import('./PluginButton') as unknown) as Promise<
-      React.ComponentType<PluginButtonProps & { config: SlatePlugin }>
-    >
-);
-
-const HoverButtons = ({
-  editor,
-  plugins,
-  translations,
-}: PluginButtonProps & SlateProps) => (
-  <div>
-    {plugins &&
-      plugins.map((plugin, i: number) =>
-        plugin.addHoverButton ? (
-          <PluginButton
-            translations={translations}
-            key={i}
-            config={plugin}
-            editor={editor}
-          />
-        ) : null
-      )}
-  </div>
-);
-
-const HoverButtonsContainer = (props: SlateProps) => {
-  const showHoverToolbar = useTextIsSelected();
-  const toolbarRef = useRef<HTMLDivElement>();
-  const editor = useSlate();
-  useEffect(() => {
-    const toolbar = toolbarRef.current;
-
-    if (!showHoverToolbar) {
-      return;
-    }
-
-    let s = window.getSelection();
-    let oRange = s.getRangeAt(0); // get the text range
-    let oRect = oRange.getBoundingClientRect();
-    if (oRect) {
-      const { left, top, width } = oRect;
-
-      toolbar.style.opacity = '1';
-      toolbar.style.top = `${top + window.scrollY - toolbar.offsetHeight}px`;
-      toolbar.style.left = `${left +
-        window.scrollX -
-        toolbar.offsetWidth / 2 +
-        width / 2}px`;
-    }
-  }, [editor, showHoverToolbar]);
-
-  return (
-    <Portal>
-      <div
-        className={
-          'ory-plugins-content-slate-inline-toolbar ' +
-          (showHoverToolbar
-            ? ''
-            : 'ory-plugins-content-slate-inline-toolbar--hidden')
-        }
-        style={{ padding: 0 }}
-        ref={toolbarRef}
-      >
-        <HoverButtons editor={editor} {...props} />
-      </div>
-    </Portal>
-  );
-};
-
-const ToolbarButtons = ({
-  plugins,
-
-  editor,
-  translations,
-}: PluginButtonProps & SlateProps) => (
-  <div>
-    {plugins &&
-      plugins.map((plugin, i: number) =>
-        plugin.addToolbarButton ? (
-          <PluginButton
-            translations={translations}
-            key={i}
-            config={plugin}
-            editor={editor}
-          />
-        ) : null
-      )}
-  </div>
-);
+const HoverButtons = lazyLoad(() => import('./HoverButtons'));
+const Toolbar = lazyLoad(() => import('./Toolbar'));
 
 const useComponentNodePlugins = (
   { plugins }: { plugins: SlatePlugin[] },
@@ -263,16 +168,19 @@ const SlateEditable = React.memo(
   ({
     plugins,
     defaultPluginType,
+    readOnly,
   }: {
     plugins: SlatePlugin[];
     defaultPluginType: string;
+    readOnly: boolean;
   }) => {
     const renderElement = useRenderElement({ plugins, defaultPluginType }, []);
     const renderLeaf = useRenderLeave({ plugins }, []);
-    const onKeyDown = useOnKeyDown({ plugins }, []);
+    const onKeyDown = readOnly ? undefined : useOnKeyDown({ plugins }, []);
 
     return (
       <Editable
+        readOnly={readOnly}
         renderElement={renderElement}
         renderLeaf={renderLeaf}
         onKeyDown={onKeyDown}
@@ -281,56 +189,8 @@ const SlateEditable = React.memo(
   }
 );
 
-const useTextIsSelected = () => {
-  const editor = useSlate();
-  try {
-    console.log(
-      Boolean(editor.selection) && Editor.string(editor, editor.selection)
-    );
-    return (
-      Boolean(editor.selection) &&
-      Editor.string(editor, editor.selection) !== ''
-    );
-  } catch (e) {
-    // can in some cases throw currently
-    return false;
-  }
-};
-const withInline = (plugins: SlatePlugin[]) => (editor: Editor) => {
-  const { isInline } = editor;
-  editor.isInline = element => {
-    return plugins.some(
-      plugin =>
-        plugin.pluginType === 'component' &&
-        plugin.object === 'inline' &&
-        plugin.type === element.type
-    )
-      ? true
-      : isInline(element);
-  };
-  return editor;
-};
-
-const withPaste = (plugins: SlatePlugin[]) => (editor: ReactEditor) => {
-  const { insertData } = editor;
-  const htmlToSlate = HtmlToSlate({ plugins });
-  editor.insertData = data => {
-    const html = data.getData('text/html');
-
-    if (html) {
-      const { slate } = htmlToSlate(html);
-
-      Transforms.insertFragment(editor, slate);
-      return;
-    }
-
-    insertData(data);
-  };
-  return editor;
-};
-
 const SlateControls = (props: SlateProps) => {
-  const { plugins, focused, readOnly, remove, translations } = props;
+  const { plugins, focused, readOnly, remove } = props;
   const editor = useMemo(
     () => withPaste(plugins)(withReact(withInline(plugins)(createEditor()))),
     []
@@ -358,26 +218,20 @@ const SlateControls = (props: SlateProps) => {
 
   return (
     <Slate editor={editor} value={value} onChange={onChange}>
-      {!readOnly && focused && <HoverButtonsContainer {...props} />}
+      {!readOnly && focused && <HoverButtons {...props} />}
 
       <SlateEditable
+        readOnly={readOnly}
         plugins={plugins}
         defaultPluginType={props.defaultPluginType}
       />
-
       {!readOnly ? (
-        <BottomToolbar
-          open={showBottomToolbar}
-          dark={true}
-          onDelete={remove}
+        <Toolbar
+          plugins={plugins}
+          show={showBottomToolbar}
+          removeSlate={remove}
           {...props}
-        >
-          <ToolbarButtons
-            {...props}
-            translations={translations}
-            editor={editor.current}
-          />
-        </BottomToolbar>
+        />
       ) : null}
     </Slate>
   );
